Tphong be

+++ services/post-service/src/main/java/org/nexo/postservice/service/impl/PostServiceImpl.java
@@ -12,7 +12,11 @@ import org.nexo.postservice.repository.IReelRepository;
 import org.nexo.postservice.service.IHashTagService;
 import org.nexo.postservice.service.IPostService;
 import org.nexo.postservice.util.Enum.EVisibilityPost;
+import org.nexo.postservice.util.Enum.SecurityUtil;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -23,12 +27,17 @@ import java.util.List;
 public class PostServiceImpl implements IPostService {
     private final IPostRepository postRepository;
     private final IReelRepository reelRepository;
+
+
     private final IPostMediaRepository postMediaRepository;
     private final AsyncFileService fileServiceClient;
     private final IHashTagService hashTagService;
+    private final SecurityUtil securityUtil;
 
     @Override
     public String savePost(PostRequestDTO postRequestDTO, List<MultipartFile> files) {
+        securityUtil.checkOwner(postRequestDTO.getUserId());
+
         PostModel model;
         if (postRequestDTO.getPostId() != 0) {
             List<PostMediaModel> postMediaModelList = postMediaRepository.findAllByPostModel_Id(postRequestDTO.getPostId());
@@ -53,7 +62,9 @@ public class PostServiceImpl implements IPostService {
         }
         postRepository.save(model);
         if (files != null && !files.isEmpty() && !files.getFirst().isEmpty()) {
-            fileServiceClient.savePostMedia(files, model.getId());
+            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+            String token = ((JwtAuthenticationToken) auth).getToken().getTokenValue();
+            fileServiceClient.savePostMedia(files, model.getId(), token);
         }
         hashTagService.findAndAddHashTagFromCaption(model);
         return "Success";
@@ -61,6 +72,8 @@ public class PostServiceImpl implements IPostService {
 
     @Override
     public String saveReel(PostRequestDTO postRequestDTO, List<MultipartFile> files) {
+        securityUtil.checkOwner(postRequestDTO.getUserId());
+
         ReelModel model;
         if (postRequestDTO.getPostId() != 0) {
             model = reelRepository.findById(postRequestDTO.getPostId())
@@ -78,7 +91,9 @@ public class PostServiceImpl implements IPostService {
         }
         reelRepository.save(model);
         if (files != null && !files.isEmpty() && !files.getFirst().isEmpty()) {
-            fileServiceClient.saveReelMedia(files, model.getId());
+            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+            String token = ((JwtAuthenticationToken) auth).getToken().getTokenValue();
+            fileServiceClient.saveReelMedia(files, model.getId(), token);
         }
         hashTagService.findAndAddHashTagFromCaption(model);
         return "Success";
@@ -86,7 +101,10 @@ public class PostServiceImpl implements IPostService {
 
     @Override
     public String inactivePost(Long id) {
+
         PostModel model = postRepository.findById(id).orElseThrow(() -> new CustomException("Post is not  exist", HttpStatus.BAD_REQUEST));
+        securityUtil.checkOwner(model.getUserId());
+
         model.setIsActive(!model.getIsActive());
         postRepository.save(model);
         return "Success";
@@ -95,6 +113,7 @@ public class PostServiceImpl implements IPostService {
     @Override
     public String inactiveReel(Long id) {
         ReelModel model = reelRepository.findById(id).orElseThrow(() -> new CustomException("Reel is not  exist", HttpStatus.BAD_REQUEST));
+        securityUtil.checkOwner(model.getUserId());
         model.setIsActive(!model.getIsActive());
         reelRepository.save(model);
         return "Success";
@@ -103,9 +122,16 @@ public class PostServiceImpl implements IPostService {
     @Override
     public String deletePost(Long id) {
         PostModel model = postRepository.findById(id).orElseThrow(() -> new CustomException("Post is not  exist", HttpStatus.BAD_REQUEST));
+        securityUtil.checkOwner(model.getUserId());
         postRepository.delete(model);
         return "Success";
     }
 
-
+    @Override
+    public String deleteReel(Long id) {
+        ReelModel model = reelRepository.findById(id).orElseThrow(() -> new CustomException("Reel is not  exist", HttpStatus.BAD_REQUEST));
+        securityUtil.checkOwner(model.getUserId());
+        reelRepository.delete(model);
+        return "Success";
+    }
 }
